#864 Moved ESC Charset Conversion into TEscCharsetConverter
